Script to filter GLIFT output based on well name.
The script can filter the output from flow when running gas lift
optimization in debug mode. The script extracts information (lines)
related to a particular well based on a regular expression.

You can use the script if you first pipe the output from flow to
a log file, for example in the terminal window:

$ flow SPE1CASE1.DATA | tee flow.log

Then after simulation has finished (or while the simulation is running):

$ filter.pl C-1H flow.log --regex=2